fix(catalog,query): SQL correctness, concurrency, and query-cache safety
Concurrency: per-query goroutine-keyed overlay for CTE/derived-table
results replaces unsynchronized writes to the shared cteResults map
under RLock (concurrent SELECTs could crash the process or contaminate
each other's results); cloneTableDef reads AutoIncSeq atomically.

Semantics: WHERE and projection evaluation errors now propagate instead
of returning empty results/NULL (SELECT 1/0 errors; a typo'd column is
an error, not an empty set); CASE/COALESCE/IIF evaluate lazily so
guarded expressions never fire; two string operands always compare as
strings ('0123' = '123' is false); the parser rejects trailing garbage
and CREATE VIEW ... UNION instead of silently truncating; indexed point
lookups coerce the search value to the column type (results no longer
depend on index existence) and integer PK keys stay exact above 2^53;
SUM preserves int64 precision; window PARTITION BY keys are
collision-safe; LAG/LEAD defaults evaluate per row; LIKE is rune-aware
with O(n*m) matching; CAST is unified on one implementation.

Query cache: dependencies extracted recursively (subqueries, derived
tables, views resolved to base tables); non-deterministic functions
rejected in every clause; unknown expression nodes are uncacheable
instead of pointer-serialized; rollback invalidates touched tables;
TTL defaults to 5 minutes; an invalidation epoch prevents re-inserting
stale rows after an unlocked scan.

Perf: per-catalog optimizer singleton; join reorder restricted to safe
consecutive INNER joins; parallel scan no longer decodes rows twice and
surfaces worker errors; ORDER BY resolves columns once and sorts with
precomputed keys. Dead evaluators removed.

+++ pkg/catalog/z_catalog_unit_coverage2_test.go
@@ -500,7 +500,7 @@ func TestUnitEvalExpression_Functions(t *testing.T) {
 		{"ABS negative", "ABS", []query.Expression{&query.NumberLiteral{Value: -5}}, 5.0},
 		{"UPPER", "UPPER", []query.Expression{&query.StringLiteral{Value: "hello"}}, "HELLO"},
 		{"LOWER", "LOWER", []query.Expression{&query.StringLiteral{Value: "HELLO"}}, "hello"},
-		{"LENGTH", "LENGTH", []query.Expression{&query.StringLiteral{Value: "hello"}}, 5},
+		{"LENGTH", "LENGTH", []query.Expression{&query.StringLiteral{Value: "hello"}}, int64(5)},
 		{"TRIM", "TRIM", []query.Expression{&query.StringLiteral{Value: "  hi  "}}, "hi"},
 		{"LTRIM", "LTRIM", []query.Expression{&query.StringLiteral{Value: "  hi"}}, "hi"},
 		{"RTRIM", "RTRIM", []query.Expression{&query.StringLiteral{Value: "hi  "}}, "hi"},
